# Improve loading performance of large SVG files with rectangular clip paths