- The SunGraphics2D implementation is more efficient when a Rectangle is used for the clip path
- Reducing the number of Graphics2D objects created aids with performance

This is in relation to #49 

For the changes in `ShapeRenderer`, there are more sophisticated ways to do it, but this seems to do the trick. The `Graphics2D` object passed is also modified in terms of clip, but this is already reset afterwards, and translate, but this should be handled by resetting the transform.

I created a repository where I did some tests, comparing the rendering time of some sample SVGs I found, before and after my changes: 

https://github.com/eriklumme1/jsvg-performance-test/blob/main/src/test/java/JsvgRenderTest.java
SVGs can be found in the test resources folder.

| File | Render before (ms) | Render after (ms) |
| --- | --- | --- |
| android.svg | 2.1 | 1.8 |
| gallardo.svg | 256 | 283 |
| islands.svg | 189 | 187 |
| micelle.svg | 270 | 233 |
| pencil.svg | 4.5 | 4.6 |
| propane.svg | 16  | 13  |
| venus.svg | 0.22 | 0.25 |
| world.svg | 4461 | 361 |

Performance is mostly the same, but the `world.svg` showed similar gains to the file that prompted me to create the issue.

I also did checksums of the generated PNG's before and after, they are different in two files. The first one is the gallardo.svg, where it looks different on the top side of the bumper. This is the only place where there are differences.

Before changes
![old-gallardo](https://github.com/weisJ/jsvg/assets/141633336/a2d1d144-3c05-44a6-a9de-8164b07e4e63)
After changes
![new-gallardo](https://github.com/weisJ/jsvg/assets/141633336/52b74f5a-7feb-4834-91e7-df1b83ceb407)
SVG in Chrome
![chrome-gallardo](https://github.com/weisJ/jsvg/assets/141633336/1240a6e8-8208-481b-9e9d-789cf2bf994d)

The other one is the world.svg, where only a bottom border is displayed differently. This might be an improvement.

Before changes
![old-world](https://github.com/weisJ/jsvg/assets/141633336/8cec6f34-957b-4bfd-97a0-c87f790563b2)
After changes
![new-world](https://github.com/weisJ/jsvg/assets/141633336/3638ae33-c16c-43da-bec7-6311daa45285)
SVG in Chrome
![chrome-world](https://github.com/weisJ/jsvg/assets/141633336/b4fa5b41-5953-4493-b62f-f7fc8f222f95)

I also had to remove one SVG that did not render with either version, although it rendered with 1.0.0. I'll make an issue for that.

Please let me know any concerns you have.